group code into the same files, rename and delete
