+++ x/government/client/cli/query.go
@@ -5,8 +5,11 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"context"
+
 	"github.com/commercionetwork/commercionetwork/x/government/types"
 	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
 )
 
 // GetQueryCmd returns the cli query commands for this module
@@ -23,5 +26,33 @@ func GetQueryCmd(queryRoute string) *cobra.Command {
 	cmd.AddCommand(
 		CmdGetGovernmentAddr(),
 	)
+	return cmd
+}
+
+func CmdGetGovernmentAddr() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "gov-address",
+		Short: "Get the government address",
+		Args:  cobra.ExactArgs(0),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			params := &types.QueryGovernmentAddrRequest{}
+
+			res, err := queryClient.GovernmentAddr(context.Background(), params)
+
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
 	return cmd
 }
