+++ x/government/keeper/keeper.go
@@ -5,6 +5,8 @@ import (
 
 	"github.com/tendermint/tendermint/libs/log"
 
+	"errors"
+
 	"github.com/commercionetwork/commercionetwork/x/government/types"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -35,7 +37,25 @@ func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 	return ctx.Logger().With("module", fmt.Sprintf("x/%s", types.ModuleName))
 }
 
-// GovernmentAddress do nothing
-func (k Keeper) GovernmentAddress(ctx sdk.Context) error {
+// SetGovernmentAddress allows to set the given address as the one that
+// the government will use later
+func (k Keeper) SetGovernmentAddress(ctx sdk.Context, address sdk.AccAddress) error {
+	if address == nil {
+		return errors.New("address is not defined")
+	}
+
+	store := ctx.KVStore(k.storeKey)
+
+	if store.Has([]byte(types.GovernmentStoreKey)) {
+		return errors.New("government address already set")
+	}
+
+	store.Set([]byte(types.GovernmentStoreKey), address)
 	return nil
+}
+
+// GetGovernmentAddress returns the address that the government has currently
+func (k Keeper) GetGovernmentAddress(ctx sdk.Context) sdk.AccAddress {
+	store := ctx.KVStore(k.storeKey)
+	return store.Get([]byte(types.GovernmentStoreKey))
 }
