+++ x/government/keeper/msg_server.go
@@ -1,8 +1,12 @@
 package keeper
 
 import (
+	"context"
+
 	"github.com/commercionetwork/commercionetwork/x/government/types"
 )
+
+var _ types.MsgServer = msgServer{}
 
 type msgServer struct {
 	Keeper
@@ -14,4 +18,7 @@ func NewMsgServerImpl(keeper Keeper) types.MsgServer {
 	return &msgServer{Keeper: keeper}
 }
 
-var _ types.MsgServer = msgServer{}
+func (k msgServer) SetGovAddress(goCtx context.Context, msg *types.MsgSetGovAddress) (*types.MsgSetGovAddressResponse, error) {
+
+	return &types.MsgSetGovAddressResponse{}, nil
+}
